fixes according to feedback
