+++ apps/property-tree/src/features/inspections/ui/mobile/MobileInspectionSheet.tsx
@@ -1,11 +1,13 @@
-import type { InspectionSubmitData } from '@/features/inspections/types/index'
+import type {
+  InspectionSubmitData,
+  TenantInfoCardData,
+} from '@/features/inspections/types/index'
 
 import type { components } from '@/services/api/core/generated/api-types'
 import type { Room } from '@/services/types'
 
 import { Sheet, SheetContent } from '@/shared/ui/Sheet'
 
-import type { TenantInfoCardData } from '../TenantInfoCard'
 import { MobileInspectionForm } from './MobileInspectionForm'
 type Inspection = components['schemas']['InternalInspection']
 type InspectionRoom = components['schemas']['InspectionRoom']
